#3.3 dodata autentifikacija na beku i #3.6 ispis da vikendicu nije moguce menjati jer je rezervisana

+++ IzdajMe/src/main/java/com/izdajMe/izdajMe/controller/HotOfferController.java
@@ -3,12 +3,15 @@ package com.izdajMe.izdajMe.controller;
 
 import com.izdajMe.izdajMe.model.Cottage;
 import com.izdajMe.izdajMe.model.HotOffer;
+import com.izdajMe.izdajMe.model.User;
 import com.izdajMe.izdajMe.services.CottageService;
 import com.izdajMe.izdajMe.services.HotOfferService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import javax.servlet.http.HttpServletRequest;
 
 @RestController
 @CrossOrigin(origins = "http://localhost:4200")
@@ -17,16 +20,33 @@ public class HotOfferController {
     private HotOfferService hotOfferService ;
 
     @PostMapping("/hotOffers/saveHotOffer")
-    public ResponseEntity<Void> saveHotOffer(@RequestBody HotOffer hotOffer){
-        if(hotOfferService.saveHotOffer(hotOffer)){
-            return ResponseEntity.ok(null);
+    public ResponseEntity<Void> saveHotOffer(@RequestBody HotOffer hotOffer, HttpServletRequest request){
+        if (request.getSession(false).getAttribute("role")!=null) {
+            if (request.getSession(false).getAttribute("role") == User.Role.cottageAdvertiser) {
+                if (hotOfferService.saveHotOffer(hotOffer)) {
+                    return ResponseEntity.ok(null);
+                } else {
+                    return new ResponseEntity<>(null, HttpStatus.NOT_ACCEPTABLE);
+                }
+            } else {
+                return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+            }
         }
-        else {
-           return new ResponseEntity<>(null, HttpStatus.NOT_ACCEPTABLE);
+        else{
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
         }
     }
     @PostMapping("/hotOffers/removeHotOffer")
-    public ResponseEntity<Boolean> removeHotOffer(@RequestBody Long id){
-        return new ResponseEntity<Boolean>(hotOfferService.removeHotOffer(id), HttpStatus.OK);
+    public ResponseEntity<Boolean> removeHotOffer(@RequestBody Long id,HttpServletRequest request){
+        if (request.getSession(false).getAttribute("role")!=null) {
+            if (request.getSession(false).getAttribute("role") == User.Role.cottageAdvertiser) {
+                return new ResponseEntity<Boolean>(hotOfferService.removeHotOffer(id), HttpStatus.OK);
+            } else {
+                return new ResponseEntity<Boolean>(false, HttpStatus.UNAUTHORIZED);
+            }
+        }
+        else{
+            return new ResponseEntity<Boolean>(false, HttpStatus.UNAUTHORIZED);
+        }
     }
 }
